requirements.txt
APScheduler==3.6.3
asgiref==3.3.1
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.11.8
cffi==1.14.3
charset-normalizer==2.0.1
click==8.0.1
colorama==0.4.4
cryptography==3.2.1
cycler==0.10.0
decorator==4.4.2
Django==3.1.4
Flask==2.0.1
gunicorn==20.1.0
idna==3.2
iniconfig==1.1.1
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.3.3
mysql-connector-python==8.0.22
numpy==1.21.0
packaging==20.9
pandas==1.3.0
Pillow==8.0.1
pluggy==0.13.1
protobuf==3.14.0
py==1.10.0
pycparser==2.20
pyparsing==2.4.7
PyRECONSTRUCT==2.2.0
pytest==6.2.2
python-dateutil==2.8.1
python-telegram-bot==13.0
pytz==2020.4
requests==2.26.0
scikit-learn==0.24.2
scipy==1.5.4
six==1.15.0
sqlparse==0.4.1
textparser==0.23.0
threadpoolctl==2.2.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.6
Werkzeug==2.0.1

PyGithub~=1.55